[ADD] Movie detail screen

+++ app/src/main/java/com/escorp/movieworld/ui/detailScreen/actorDetail/actorsCredits/ActorCreditsViewModel.kt
@@ -6,5 +6,5 @@ import javax.inject.Inject
 
 class ActorCreditsViewModel @Inject constructor(private val dataRepository: DataRepository) : ViewModel() {
 
-    fun getPersonsCredits(personId: Long) = dataRepository.getPersonCredits(personId)
+    fun getPersonsCredits(personId: Int) = dataRepository.getPersonCredits(personId)
 }
